+++ spec/support/action_controller_spec_helper.rb
@@ -1,4 +1,3 @@
 require "ostruct"
 require "active_record"
 require "action_controller"
-require "active_support/configurable"