+++ spec/support/active_record_spec_helper.rb
@@ -1,2 +1,1 @@
-require "active_support/configurable"
 require "active_record"